Clean up NotesProvider further.
