+++ src/components/NotesProvider/consts/labels.ts
@@ -1,0 +1,3 @@
+export const LABEL_LOCAL = "local";
+export const LABEL_REMOTE = "remote";
+export const LABEL_IMPORTED = "imported:";
